refactor(activerecord): extract Base.enum macro body to enum.ts
Moves the ~100-line implementation of `Base.enum(attribute, mapping, options)`
out of base.ts into enum.ts as a `this`-typed function (exported as
`enumMethod` internally + aliased to `enum` via `export { enumMethod as enum }`
so the extractor finds it under Rails' `enum` name), wired onto Base via
extend(). The binding stays at `Base.enum(...)` — same user-facing API,
now living next to EnumType / defineEnum / readEnumValue where Rails'
enum.rb keeps it.

No behavior change. Preserves the in-memory-only bang setter semantics
that differentiate `Base.enum` from the fuller `defineEnum` (the latter
is the Rails-persisting variant with not* scopes).

api:compare:
  moves 209 → 208 (−1, the `enum` move is resolved)
  inheritance 199/210 → 200/210 (95.2%)
  matched 2514/2819 (unchanged)

17793 tests pass.
